Next_Read_buffer_with_LoRa

+++ LoRa_Receiv_Mic/myLoRa.h
@@ -24,5 +24,5 @@ class myLoRa {
 		byte read_pin();
 
   private:
-    static int pointer;
+    static signed long pointer;
 };
